third_party/WebKit/LayoutTests/platform/mac-mac10.10/external/wpt/IndexedDB/nested-cloning-small-expected.txt

This is a testharness.js-based test.
FAIL small typed array assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL blob assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL blob with small typed array assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL blob with small typed array with key generator assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL blob array assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL blob array with key generator assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL array of blobs and small typed arrays assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL array of blobs and small typed arrays with key generator assert_equals: Expected success event, but got error event instead expected "success" but got "error"
Harness: the test ran to completion.
